requirements.txt
Django>=4.2
djangorestframework
gunicorn
django-environ==0.12.0
black==24.3.0
flake8==6.1.0
isort==5.12.0
detect-secrets==1.4.0
pytest
pytest-django
